refactor: extract AppErrorOperation.swift from AppErrorTypes.swift
Splits the `AppErrorOperation` raw-string enum out of AppErrorTypes.swift.

Byte-identical move of lines 3-16. Both files import Foundation.

AppErrorTypes.swift: 27 → 12 lines. Tests: 667.
